Add files via upload
The design process for the YouTube Festival Awards 2022 was a vibrant and engaging endeavor. We started by conceptualizing a dynamic festival atmosphere, featuring key elements like a striking red arch-rack entrance, a spacious main stage, a smaller stage for intimate sessions, and a dedicated food area. To infuse the YouTube spirit, we incorporated colorful boxes inspired by YouTube's palette, providing seating and interactive spaces. With 500 invited guests in mind, we meticulously planned the registration space for smooth check-ins. To create an immersive experience, large festival standees adorned the venue, and a tunnel served as the entrance, linking Ballroom One to Ballroom Two. We also integrated YouTube Shorts for fun photo opportunities, ensuring that the festival design was not only visually captivating but also aligned with the event's goal of sharing insights and celebrating the platform's success and future potential with advertisers, marketers, creators, and brands.